basis.router: place tests for parsing step in separate file
